Move cli repl loop into its own file.
